Can now find (potential) references to a method
If you place the caret at a method definition you can now find all
of the potential references to the method. I say potential because
I don't currently use the compiler to validate that it's a
reference to the exact method under the cursor.

I list the references in a dialog so it's not really useful and it
seems that it might be recording some references twice.