Failed attempt
I tried to use a simple approach that would have allowed us
to show type-hiearchies that were based on concrete types rather
than just generic ones, i.e. that the sub-type hierarchy of
Foo[String] would be different than Foo[Int] depending.

However, I ouldn't find a way a way to get the full type at a
given location, i.e. I ran into the same issue as we had with
overloaded symbols

example

trait Foo[A]
trait StringFoo extends |Foo[String]

If | represents the location, then asking for the tree.tpe of
that location will return Foo rather than Foo[String]. I tried to
use the same approach as we did with overloaded symbols, i.e.
use a AppliedTypeTree (in symbols we used a TypedApply) locator
but for some reason I couldn't get it to work (it would give me
an <empty> tree even though I used the right RangePosition to
mark the entire type).

So I'm shelfing this :( the type-hierarchy will only be able to
be based on generic types (though it will be able to show the
concrete type in the list of super-types)